day8/test_main.py
from .main import part1, part2


def test_part1():
    input = """
RL

AAA = (BBB, CCC)
BBB = (DDD, EEE)
CCC = (ZZZ, GGG)
DDD = (DDD, DDD)
EEE = (EEE, EEE)
GGG = (GGG, GGG)
ZZZ = (ZZZ, ZZZ)
"""

    assert part1(input) == 2

    input = """
LLR

AAA = (BBB, BBB)
BBB = (AAA, ZZZ)
ZZZ = (ZZZ, ZZZ)
"""

    assert part1(input) == 6


def test_part2():
    input = """
LR

11A = (11B, XXX)
11B = (XXX, 11Z)
11Z = (11B, XXX)
22A = (22B, XXX)
22B = (22C, 22C)
22C = (22Z, 22Z)
22Z = (22B, 22B)
XXX = (XXX, XXX)
"""

    assert part2(input) == 6


def test_input():
    with open("input.txt") as file:
        input = file.read()

    assert part1(input) == 12599
    assert part2(input) == 8245452805243
